+++ java-client/src/main/java/org/opensearch/client/json/JsonpSerializable.java
@@ -33,6 +33,9 @@
 package org.opensearch.client.json;
 
 import jakarta.json.stream.JsonGenerator;
+import java.io.IOException;
+import java.io.StringWriter;
+import java.io.UncheckedIOException;
 
 /**
  * An object that is its own JsonP serializer
@@ -40,4 +43,24 @@ import jakarta.json.stream.JsonGenerator;
 public interface JsonpSerializable {
 
     void serialize(JsonGenerator generator, JsonpMapper mapper);
+
+    /**
+     * A default method which returns string representation for the instances of classes
+     * implementing JsonpSerializable interface.<br>
+     * Usage : Eg for SearchRequest.class<pre>{@code SearchRequest implements JsonpSerializable{..}
+     * SearchRequest searchRequest = SearchRequest.of(request -> request...);
+     * String searchRequestString = searchRequest.writeValueAsString();}</pre>  <br>
+     *
+     */
+    default String toJsonString() {
+        try (StringWriter writer = new StringWriter()) {
+            try (JsonGenerator generator = JsonpUtils.DEFAULT_PROVIDER.createGenerator(writer)) {
+                serialize(generator, JsonpUtils.DEFAULT_JSONP_MAPPER);
+            }
+            return writer.toString();
+        } catch (IOException ex) {
+            throw new UncheckedIOException(ex);
+        }
+    }
+
 }